+++ packages/tabs/src/Tabs.spec.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, fireEvent, cleanup } from '@testing-library/react';
-import 'jest-dom/extend-expect';
+import '@testing-library/jest-dom/extend-expect';
 import { Tabs, Tab } from './index';
 
 afterAll(cleanup);