+++ packages/toggle/src/Toggle.story.tsx
@@ -2,33 +2,30 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { boolean, select } from '@storybook/addon-knobs';
 import { css } from '@leafygreen-ui/emotion';
+import { uiColors } from '@leafygreen-ui/palette';
 import Toggle, { Size, Variant } from '.';
 
-const containerStyle = css`
-  padding: 2rem;
-  text-align: center;
-`;
+storiesOf('Toggle', module).add('Default', () => {
+  const variant = select(
+    'Variant',
+    Object.values(Variant) as Array<Variant>,
+    Variant.Default,
+  );
 
-storiesOf('Toggle', module)
-  .add('Default', () => (
-    <div className={containerStyle}>
-      <Toggle
-        size={select('Size', Object.values(Size) as Array<Size>, Size.Default)}
-        disabled={boolean('Disabled', false)}
-      />
-    </div>
-  ))
-  .add('Dark', () => (
+  return (
     <div
       className={css`
-        ${containerStyle};
-        background-color: #464c4f;
+        background-color: ${variant === 'default'
+          ? uiColors.gray.light1[1]
+          : uiColors.gray.dark1};
+        padding: 20px;
       `}
     >
       <Toggle
+        variant={variant}
         size={select('Size', Object.values(Size) as Array<Size>, Size.Default)}
-        variant={Variant.Dark}
         disabled={boolean('Disabled', false)}
       />
     </div>
-  ));
+  );
+});