+++ packages/tooltip/src/Tooltip.spec.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import 'jest-dom/extend-expect';
+import '@testing-library/jest-dom/extend-expect';
 import { render, fireEvent, cleanup } from '@testing-library/react';
 import Tooltip from '.';
 
@@ -241,6 +241,37 @@ describe('packages/Tooltip', () => {
 
       fireEvent.click(button);
       expect(tooltip).not.toBeVisible();
+    });
+  });
+
+  describe('when trigger contains nested children', () => {
+    interface ButtonProps {
+      children: React.ReactNode;
+    }
+
+    function Button({ children, ...props }: ButtonProps) {
+      return (
+        <button {...props} data-testid="nested-trigger">
+          trigger {children}
+        </button>
+      );
+    }
+
+    const { getByTestId } = render(
+      <Tooltip
+        trigger={
+          <Button>
+            <span>trigger</span>
+          </Button>
+        }
+      >
+        <div>Tooltip!</div>
+      </Tooltip>,
+    );
+
+    test('renders trigger in document', () => {
+      const trigger = getByTestId('nested-trigger');
+      expect(trigger).toBeInTheDocument();
     });
   });
 });