+++ packages/lib/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @leafygreen-ui/lib
 
+## 4.0.0
+
+### Major Changes
+
+- 464c09d: Introduces SSR compatibility though a change to our build process and files
+
+### Patch Changes
+
+- Updated dependencies [464c09d]
+  - @leafygreen-ui/emotion@2.0.0
+
+## 3.2.0
+
+### Minor Changes
+
+- 2f9a300: Add keyMap to lib
+
 ## 3.1.0
 
 ### Minor Changes