+++ packages/menu/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @leafygreen-ui/menu
 
+## 5.0.0
+
+### Major Changes
+
+- 464c09d: Introduces SSR compatibility though a change to our build process and files
+
+### Minor Changes
+
+- 5c9202d: Introduces support for LeafyGreenProvider for improved focus state management
+
+### Patch Changes
+
+- 31f6bfd: Initial release of @leafygreen-ui/pipeline component. Removed deprecated title prop from the @leafygreen-ui/menu component propTypes.
+- Updated dependencies [5c9202d]
+- Updated dependencies [464c09d]
+  - @leafygreen-ui/leafygreen-provider@1.0.0
+  - @leafygreen-ui/hooks@2.0.0
+  - @leafygreen-ui/lib@4.0.0
+  - @leafygreen-ui/palette@2.0.0
+  - @leafygreen-ui/popover@3.0.0
+
+## 4.0.3
+
+### Patch Changes
+
+- 8c2e838: Update styles to fix bugs on Safari and Firefox
+
+## 4.0.2
+
+### Patch Changes
+
+- ad74307: Change default behavior on MenuItems with anchor tags, such that the default is to target="\_self", and ensure ability of consumer to override the default
+- ff55bb5: Update anchor styles to minimize overrides
+- baf45f6: Allow target and rel props to be passed to MenuItem, when MenuItem is a link
+
+## 4.0.1
+
+### Patch Changes
+
+- 2f9a300: Use exported keyMap from lib
+- 759523f: Add border-box property to MenuItems and update documentation
+- Updated dependencies [2f9a300]
+  - @leafygreen-ui/lib@3.2.0
+
 ## 4.0.0
 
 ### Major Changes