+++ packages/palette/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @leafygreen-ui/palette
 
+## 2.0.0
+
+### Major Changes
+
+- 464c09d: Introduces SSR compatibility though a change to our build process and files
+
+## 1.1.1
+
+### Patch Changes
+
+- 319844d: Fix bug in storybook such that HEX values are now displayed
+
 ## 1.1.0
 
 ### Minor Changes