+++ packages/tabs/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @leafygreen-ui/tabs
 
+## 2.0.0
+
+### Major Changes
+
+- 464c09d: Introduces SSR compatibility though a change to our build process and files
+
+### Patch Changes
+
+- Updated dependencies [464c09d]
+  - @leafygreen-ui/lib@4.0.0
+  - @leafygreen-ui/palette@2.0.0
+
+## 1.0.5
+
+### Patch Changes
+
+- 2f9a300: Use exported keyMap from lib
+- Updated dependencies [2f9a300]
+  - @leafygreen-ui/lib@3.2.0
+
 ## 1.0.4
 
 ### Patch Changes