+++ packages/toggle/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @leafygreen-ui/toggle
 
+## 3.0.0
+
+### Major Changes
+
+- 464c09d: Introduces SSR compatibility though a change to our build process and files
+
+### Patch Changes
+
+- Updated dependencies [464c09d]
+  - @leafygreen-ui/lib@4.0.0
+  - @leafygreen-ui/palette@2.0.0
+
+## 2.1.8
+
+### Patch Changes
+
+- 0868af7: Update Toggle to new utilize new palette
+
 ## 2.1.7
 
 ### Patch Changes