+++ packages/tooltip/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @leafygreen-ui/tooltip
 
+## 3.0.0
+
+### Major Changes
+
+- 464c09d: Introduces SSR compatibility though a change to our build process and files
+
+### Patch Changes
+
+- Updated dependencies [464c09d]
+  - @leafygreen-ui/hooks@2.0.0
+  - @leafygreen-ui/lib@4.0.0
+  - @leafygreen-ui/palette@2.0.0
+  - @leafygreen-ui/popover@3.0.0
+
+## 2.0.2
+
+### Patch Changes
+
+- 13e3eab: Enable trigger prop to accept nested components
+- Updated dependencies [319844d]
+  - @leafygreen-ui/palette@1.1.1
+
+## 2.0.1
+
+### Patch Changes
+
+- 50853ca: Upgrade dependencies
+
 ## 2.0.0
 
 ### Major Changes